backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai>=1.52.0
aiortc>=1.9.0
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
requests==2.31.0
numpy>=1.24.3
sounddevice==0.4.6
edge-tts==6.1.1
faster-whisper>=1.0.0
jiwer>=3.0.0
aiosqlite>=0.19.0
yt-dlp>=2024.1.0
python-multipart>=0.0.9


